Removed firebase from examples
