+++ package.json
@@ -22,9 +22,7 @@
     "@angular/platform-browser": "^6.0.0",
     "@angular/platform-browser-dynamic": "^6.0.0",
     "@angular/router": "^6.0.0",
-    "angularfire2": "^5.0.0-rc.11",
     "core-js": "^2.5.4",
-    "firebase": "^5.1.0",
     "gsap": "^2.0.1",
     "jquery": "^3.3.1",
     "rxjs": "^6.0.0",
