--- src/app/examples/09-slideshow/images.ts
@@ -1,5 +1,0 @@
-export const images: string[] = [
-  'assets/lion-roar.jpg',
-  'assets/maxres.jpg',
-  'assets/maxresdefault.jpg'
-];
