+++ src/app/examples/09-slideshow/slideshow.component.ts
@@ -1,29 +1,57 @@
-import { Component } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { fromEvent, merge } from 'rxjs/index';
+import { map, scan, startWith } from 'rxjs/operators';
+import { slideshowAnimation } from './slideshow.animations';
+
+const images: string[] = [
+  'assets/lion-roar.jpg',
+  'assets/maxres.jpg',
+  'assets/maxresdefault.jpg'
+];
 
 @Component({
   selector: 'app-slideshow',
-  styles: [`
-  mat-card {
-    width: 100%;
-    max-width: 900px;
-    box-sizing: border-box;
-    margin: 16px;
-    overflow: hidden;
-  }
-  .card-container {
-    display: flex;
-    flex-flow: row wrap;
-  }
-  `],
-  template: `
-  <div class="card-container">
-      <mat-card>
-        <app-slideshow-master></app-slideshow-master>
-      </mat-card>
-      <mat-card>
-        <app-slideshow-client></app-slideshow-client>
-      </mat-card>
-  </div>
-  `
+  styleUrls: ['./slideshow.component.css'],
+  templateUrl: './slideshow.component.html',
+  animations: [slideshowAnimation]
 })
-export class SlideshowComponent {}
+export class SlideshowComponent implements OnInit {
+  @ViewChild('previous') previous;
+  @ViewChild('next') next;
+  images: any[] = images;
+  position: any;
+  currentIndex = 0;
+  currentDirection = 'left';
+
+  constructor() {}
+
+  ngOnInit() {
+    const previous$ = fromEvent(this.getNativeElement(this.previous), 'click')
+      .pipe(map(event => ({shift: -1, direction: 'right'})));
+
+    const next$ = fromEvent(this.getNativeElement(this.next), 'click')
+      .pipe(map(event => ({shift: +1, direction: 'left'})));
+
+    merge(previous$, next$)
+      .pipe(
+        startWith({index: 0} as any),
+        scan((acc, curr) => {
+          const projectedIndex = acc.index + curr.shift;
+
+          const adjustedIndex = projectedIndex < 0 ? this.images.length - 1
+            : projectedIndex >= this.images.length ? 0
+              : projectedIndex;
+
+          return {index: adjustedIndex, direction: curr.direction};
+        })
+      )
+      .subscribe(event => {
+        this.currentIndex = event.index;
+        this.currentDirection = event.direction;
+      });
+  }
+
+  getNativeElement(element) {
+    return element._elementRef.nativeElement;
+  }
+}
