+++ src/app/examples/index.ts
@@ -8,33 +8,19 @@ export { SimpleAnimationComponent } from './06-simple-animation/simple-animation
 export { AnimationComponent } from './07-animation/animation.component';
 
 export { CounterComponent } from './08-counter/counter.component';
-export { CounterClientComponent } from './08-counter/counter-client.component';
-export { CounterMasterComponent } from './08-counter/counter-master.component';
 
 export { SlideshowComponent } from './09-slideshow/slideshow.component';
-export { SlideshowClientComponent } from './09-slideshow/slideshow-client.component';
-export { SlideshowMasterComponent } from './09-slideshow/slideshow-master.component';
 
 export { LocationComponent } from './10-location/location.component';
-export { LocationClientComponent } from './10-location/location-client.component';
-export { LocationMasterComponent } from './10-location/location-master.component';
 
 export { MapComponent } from './11-map/map.component';
-export { MapClientComponent } from './11-map/map-client.component';
-export { MapMasterComponent } from './11-map/map-master.component';
 
 export { AnnotateComponent } from './12-annotate/annotate.component';
-export { AnnotateClientComponent } from './12-annotate/annotate-client.component';
-export { AnnotateMasterComponent } from './12-annotate/annotate-master.component';
 export { DocComponent } from './12-annotate/doc.component';
 
 export { GameComponent } from './13-game/game.component';
-export { GameClientComponent } from './13-game/game-client.component';
-export { GameMasterComponent } from './13-game/game-master.component';
 
 export { SliderComponent } from './14-slider/slider.component';
-export { SliderClientComponent } from './14-slider/slider-client.component';
-export { SliderMasterComponent } from './14-slider/slider-master.component';
 
 export { WaitForStreamComponent } from './15-wait-for-stream/wait-for-stream.component';
 export { CombiningStreamsComponent } from './16-combining-streams/combining-streams.component';
