--- src/app/firebase.conf.example.ts
@@ -1,6 +1,0 @@
-export const firebaseConfig = {
-  apiKey: '<YOUR_API_KEY>',
-  authDomain: '<YOUR_AUTH_DOMAIN>',
-  databaseURL: '<YOUR_DATABASE_URL>',
-  storageBucket: '<OPTIONAL>'
-};
