--- src/app/firebase.conf.ts
@@ -1,6 +1,0 @@
-export const firebaseConfig = {
-  apiKey: 'AIzaSyD2QZmb4zNDhsV58DHe_I-V2RzgOIXerPg',
-  authDomain: 'hello-angularfire2.firebaseapp.com',
-  databaseURL: 'https://hello-angularfire2.firebaseio.com',
-  storageBucket: ''
-};
